requirements.txt
poke-env>=0.6.0
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-chroma>=0.0.1
langchain-community>=0.0.1
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.18
numpy>=1.24.0
pandas>=2.0.0 